silence compiler warnings

+++ oniguruma/enc/euc_jp.c
@@ -115,8 +115,8 @@ code_to_mbc(OnigCodePoint code, UChar *buf)
 #if 1
   if (enclen(ONIG_ENCODING_EUC_JP, buf) != (p - buf))
     return ONIGERR_INVALID_CODE_POINT_VALUE;
-#endif  
-  return p - buf;
+#endif
+  return (int)(p - buf);
 }
 
 static int
