+++ sftp/vis.c
@@ -156,7 +156,7 @@ strvis(char *dst, const char *src, int flag)
 	for (start = dst; (c = *src);)
 		dst = vis(dst, c, flag, *++src);
 	*dst = '\0';
-	return (dst - start);
+	return (int)(dst - start);
 }
 
 int
@@ -183,7 +183,7 @@ strnvis(char *dst, const char *src, size_t siz, int flag)
 			}
 			src++;
 		} else {
-			i = vis(tbuf, c, flag, *++src) - tbuf;
+			i = (int)(vis(tbuf, c, flag, *++src) - tbuf);
 			if (dst + i <= end) {
 				memcpy(dst, tbuf, i);
 				dst += i;
@@ -200,7 +200,7 @@ strnvis(char *dst, const char *src, size_t siz, int flag)
 		while ((c = *src))
 			dst += vis(tbuf, c, flag, *++src) - tbuf;
 	}
-	return (dst - start);
+	return (int)(dst - start);
 }
 
 int
@@ -216,5 +216,5 @@ strvisx(char *dst, const char *src, size_t len, int flag)
 	if (len)
 		dst = vis(dst, *src, flag, '\0');
 	*dst = '\0';
-	return (dst - start);
+	return (int)(dst - start);
 }
